# Improve dump_integer performance