This PR improves the performance of `dump_integer` using a tuned int2ascii implementation. The technique is discussed in [Fastware talk by Andrei Alexandrescu](https://youtu.be/o4-CwDo2zpg?t=1680) and has been already incorporated in some of other well-known libraries like [fmtlib](https://github.com/fmtlib/fmt/blob/master/include/fmt/format.h#L954).

In this pull request:

1. The fast int2ascii implementation is implemented (adapted from fmtlib implementation)
2. A benchmark for small integers is added. Previous benchmarks for integer are biased to very large integers. Also, in the original talk Andrei argues that the performance gain is larger (~2x) for smaller integers compared to a Naive implementation. Therefore, I believe it is worth covering the small integers in a separate benchmark too since in most use cases in the real world the integers are not that large. This lets us to keep track of the performance for both small and larger integer values.
3. I couldn't find unit-tests for integer formatting that checks different number of digits. To bring the code coverage to 100%, I have added some unit tests too.

The benchmarks show an considerable performance improvement (~1.5x) on my VM linux. This could be different (and probably higher) in a non-VM environment. Here is the original performance of Dump* operations (Ubuntu 18.04 x64, Clang 6):

```
Run on (4 X 2601 MHz CPU s)
CPU Caches:
  L1 Data 32K (x1)
  L1 Instruction 32K (x1)
  L2 Unified 256K (x1)
  L3 Unified 6144K (x1)
Dump/jeopardy / -           310113722 ns  308222627 ns         45   162.468MB/s
Dump/jeopardy / 4           367662837 ns  367510814 ns         38   181.292MB/s
Dump/canada / -              13027324 ns   13027200 ns       1096   153.024MB/s
Dump/canada / 4              16393044 ns   16328034 ns        858   473.766MB/s
Dump/citm_catalog / -         1956277 ns    1951371 ns       7267   244.506MB/s
Dump/citm_catalog / 4         2648083 ns    2647870 ns       5007   622.081MB/s
Dump/twitter / -              1912959 ns    1912949 ns       7319    232.77MB/s
Dump/twitter / 4              2181181 ns    2171835 ns       6523   336.927MB/s
Dump/floats / -             107072183 ns  106987779 ns        128    166.43MB/s
Dump/floats / 4             117135279 ns  117133075 ns        119   192.724MB/s
Dump/signed_ints / -         72700522 ns   72323217 ns        192   268.724MB/s
Dump/signed_ints / 4         80542202 ns   80504633 ns        173   300.645MB/s
Dump/unsigned_ints / -       61854487 ns   61853677 ns        230   314.503MB/s
Dump/unsigned_ints / 4       69747684 ns   69491170 ns        202   348.555MB/s
Dump/small_signed_ints / -   31521619 ns   31516332 ns        438   223.579MB/s
Dump/small_signed_ints / 4   39854903 ns   39854301 ns        349   296.449MB/s
```

And this is the performance with fast int2ascii method (note the last six benchmarks):

```
Dump/jeopardy / -           286181485 ns  286013895 ns         48   175.083MB/s
Dump/jeopardy / 4           349974284 ns  349965305 ns         40   190.381MB/s
Dump/canada / -              13124489 ns   13047201 ns       1069   152.789MB/s
Dump/canada / 4              17334759 ns   17313192 ns        876   446.808MB/s
Dump/citm_catalog / -         1871325 ns    1871283 ns       7435   254.971MB/s
Dump/citm_catalog / 4         2462448 ns    2462047 ns       5452   669.033MB/s
Dump/twitter / -              1825330 ns    1817293 ns       7690   245.022MB/s
Dump/twitter / 4              2058264 ns    2057411 ns       6820   355.666MB/s
Dump/floats / -             107943518 ns  107941378 ns        129    164.96MB/s
Dump/floats / 4             118865187 ns  118097256 ns        117   191.151MB/s
Dump/signed_ints / -         42482922 ns   42401566 ns        329   458.355MB/s
Dump/signed_ints / 4         49180866 ns   49179644 ns        285   492.142MB/s
Dump/unsigned_ints / -       41778353 ns   41575626 ns        333   467.898MB/s
Dump/unsigned_ints / 4       48426047 ns   48365002 ns        288   500.807MB/s
Dump/small_signed_ints / -   25048480 ns   25048284 ns        554   281.313MB/s
Dump/small_signed_ints / 4   34388632 ns   34332533 ns        401   344.128MB/s
````

More detailed benchmark results (including gcc, msvc) is available here: https://docs.google.com/spreadsheets/d/1sGE3UvLeLT_KrnC_Ujvtl_NFXPIp28xFOe9u9-6GSU0/edit?usp=sharing

This is my first contribution so any feedback is welcome.

* * *

## Pull request checklist

Read the [Contribution Guidelines](https://github.com/nlohmann/json/blob/develop/.github/CONTRIBUTING.md) for detailed information.

- [x]  Changes are described in the pull request, or an [existing issue is referenced](https://github.com/nlohmann/json/issues).
- [x]  The test suite [compiles and runs](https://github.com/nlohmann/json/blob/develop/README.md#execute-unit-tests) without error.
- [x]  [Code coverage](https://coveralls.io/github/nlohmann/json) is 100%. Test cases can be added by editing the [test suite](https://github.com/nlohmann/json/tree/develop/test/src).
- [x]  The source code is amalgamated; that is, after making changes to the sources in the `include/nlohmann` directory, run `make amalgamate` to create the single-header file `single_include/nlohmann/json.hpp`. The whole process is described [here](https://github.com/nlohmann/json/blob/develop/.github/CONTRIBUTING.md#files-to-change).

## Please don't

- The C++11 support varies between different **compilers** and versions. Please note the [list of supported compilers](https://github.com/nlohmann/json/blob/master/README.md#supported-compilers). Some compilers like GCC 4.7 (and earlier), Clang 3.3 (and earlier), or Microsoft Visual Studio 13.0 and earlier are known not to work due to missing or incomplete C++11 support. Please refrain from proposing changes that work around these compiler's limitations with `#ifdef`s or other means.
- Specifically, I am aware of compilation problems with **Microsoft Visual Studio** (there even is an [issue label](https://github.com/nlohmann/json/issues?utf8=✓&q=label%3A%22visual+studio%22+) for these kind of bugs). I understand that even in 2016, complete C++11 support isn't there yet. But please also understand that I do not want to drop features or uglify the code just to make Microsoft's sub-standard compiler happy. The past has shown that there are ways to express the functionality such that the code compiles with the most recent MSVC - unfortunately, this is not the main objective of the project.
- Please refrain from proposing changes that would **break [JSON](http://json.org) conformance**. If you propose a conformant extension of JSON to be supported by the library, please motivate this extension.
- Please do not open pull requests that address **multiple issues**.