@nickaein 2019-01-02
The Appveyor is failing with following message:

> Cannot assign job to clouds/groups 'pro, pro-backup'. All clouds within the group are either failing, or offline, or busy.

Apparently they are [experiencing some outage](https://status.appveyor.com/incidents/39tq9r92flzh) for now. Is there any way to re-run Appveyor CI for this PR after they resolve the issue? 

Its latest status is being reported here: https://status.appveyor.com/