@nlohmann 2019-01-02
Will do. Thanks for the PR! I am looking forward to having this performance improvement in the library!