@jaredgrubb 2019-01-03
This patch looks pretty good and seems reasonable.

My only concern is whether this patch would have any licensing implications. Was this patch developed from scratch or copied from somewhere? If it is copied from the video, was there anything in the video about the license that the code is released under? Just because it's on youtube doesn't necessarily make it public domain.