@nickaein 2019-01-03
I have rewritten the implementation with some optimizations along the way. The new implementation benchmark as following with `clang` compiler:

```
Dump/signed_ints / -         39328352 ns   39117135 ns        365   496.841MB/s
Dump/signed_ints / 4         46257733 ns   45992578 ns        299   526.245MB/s
Dump/unsigned_ints / -       36924171 ns   36911475 ns        385   527.022MB/s
Dump/unsigned_ints / 4       46389870 ns   45955745 ns        304   527.062MB/s
Dump/small_signed_ints / -   23804581 ns   23831036 ns        585   295.682MB/s
Dump/small_signed_ints / 4   34276080 ns   34128556 ns        414   346.184MB/s
```

Compared to implementation adapted from `fmtlib`, there is a ~5% performance improvement on average on benchmarks.

I've also tried this implementation with a larger lookup table with 1000 _per-calculated_ numbers (zero to 999) instead of 100 numbers (zero to 99) and it lead to a speedup of ~9% compared to `fmtlib` implementation.

Though I haven't ran benchmarks on other compilers, I hope there will be not a performance drop on them at least.

We have three options to proceed:

1. Use the current implementation in PR as it is. If there is a copyright concern on adapting code from `fmtlib` which has BSD license, this option is off the table.
2. Use the rewritten implementation which is similar to `fmtlib` implementation, but with a ~5% performance boost on `clang`. Though I have to test this implementation more to make sure it doesn't drop performance on other compilers and it also passes CI.
3. Use the rewritten implementation with bigger (1000 numbers) lookup table that gives us ~9% performance boost compared to `fmtlib` on `clang`. The tests required in option 2 must be done in this case too. More importantly, we also have to consider whether the larger lookup table (which lead to ~1KB increase in the object binary) worth its performance boost.