@nlohmann 2019-01-03
As for the copyright issue, I would like to watch the video myself. I hope to find time for this on the weekend.