@nlohmann 2019-01-03
(Any idea why the AppVeyor tests fail?)