@nickaein 2019-01-03
@nlohmann Thanks. I will wait for your response on how to proceed.

As for Appveyor, I mentioned earlier that there was downtime in their servers when I submitted the PR, hence the failure. Here is more detail: https://status.appveyor.com/incidents/39tq9r92flzh

Apparently there is no facility that automatically re-runs the CI. I wonder if it can be done manually? As workaround, I also can submit "noop" commit to trigger the CI.