@nlohmann 2019-01-03
In fact I restarted the jobs and now there are actual failures, see https://ci.appveyor.com/project/nlohmann/json/builds/21351818/job/4y0907w9i0ri4dxv.