@nlohmann 2019-01-03
(And BSD and MIT should be no problem IIRC)