@nlohmann 2019-01-06
Now I watched the video and I think there is no issue with copyrights. The video presents ideas and exemplifies them with code snippets. However, I did not find a source saying that it is OK to re-license BSD-licensed code under a MIT license.