@nickaein 2019-01-06
I didn't find any sources either. In such cases where two open-source projects have different (and perhaps incompatible) licensing, the safest approach seems to be incorporating the source license into the destination project. However, that's too much burden for us in this case.

[As mentioned eariler](https://github.com/nlohmann/json/pull/1411#issuecomment-451243254), I've already re-implemented the technique from scratch which actually has a slight performance improvement.

Is it fine to add this implementation as a commit on top of previous commits on this pull request? Or it's better to open a new pull request that drops [this commit](https://github.com/nlohmann/json/pull/1411/commits/9bd112e9ee47d7fb109065519f3656b06b221cf7) from history which in some parts contains code from fmtlib?