@nickaein 2019-01-06
I've push the new changes which are present in this commit: https://github.com/nlohmann/json/pull/1411/commits/7a6102f73955e0c0a06a8ca20003a7ee2dabce87. The other commits are untouched and same as before.