@nickaein 2019-01-11
> Why is that so? Why the function in serializer is not reused?

The only reason could be my limited knowledge of the codebase. Can you point out where is the implementation? I couldn't find such method in `serializer.hpp` except mine. Also since these methods are tuned, we might get an inferior performance with a equivalent but less performing implementation.

BTW thank you for the feedback! Please let me know if you have other suggestions.