@nlohmann 2019-01-12
I just ran the benchmarks myself and can confirm a significant performance improvement. Well done!

For the licensing part: We must not take code with a non-MIT license. This would make this library a time-bomb for future use. As long as the code was only motivated by the YouTube video, I see no problems. But anything from fmtlib should go out.

@nickaein Can you realize this?