@nickaein 2019-01-12
You're welcome! Thank you for this great library! I'm aware that the speed is not the top design goal so there could be the ways that improves the performance without affecting public API or causes a major impact on library design.

As for licensing, I understand the significance of possible license violation. The initial implementation was partially adapted from fmtlib (e.g. `static const digits` string). This choice was actually by random chance and for convenient, not because fmtlib is the only/first implementer of this technique or theirs has any particular significance.

Nevertheless, I have re-written the implementation. The current code isn't derived from fmtlib in any way (neither in its design nor in implementation details).