@nlohmann 2019-01-13
Thanks a lot!